requirements.txt
absl-py==2.4.0
annotated-types==0.7.0
astunparse==1.6.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.0.post2
charset-normalizer==3.4.4
cicflowmeter==0.5.0
click==8.3.1
cryptography==46.0.6
et_xmlfile==2.0.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
gast==0.7.0
google-auth==2.49.1
google-genai==0.5.0
google-pasta==0.2.0
grpcio==1.78.1
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.19.0
OTXv2==1.5.11
packaging==26.0
pandas==3.0.1
pillow==11.1.0
protobuf==6.33.5
psutil==5.9.8
psycopg2-binary==2.9.10
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0b3
pydantic_core==2.45.0
Pygments==2.19.2
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
qrcode==8.0
reportlab==4.3.1
requests==2.32.5
rich==14.3.3
scapy==2.6.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
six==1.17.0
SQLAlchemy==2.0.48
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
virustotal-python==1.1.1
websockets==14.2
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.1
yara-python==4.5.1
